fixing slider circular dependencies
